+++ cli.py
@@ -93,3 +93,6 @@ if not user_dumb:
     scraper = Scraper(udemy.sites)
     scrape()
 input()
+
+
+